Add `ubjectIndex()` function and update `UbjectError` structure
This commit adds a new function called ubjectIndex() to the code. This function returns the index of the given object in the object table.

It also updates the UbjectError structure to use function pointers for printing errors and warnings. This allows the user to specify their own functions for printing errors and warnings.